Struct syntaxdb::SyntaxDbClient

pub struct SyntaxDbClient { /* fields omitted */ }

Methods

impl SyntaxDbClient

fn new() -> SyntaxDbClient

fn get_languages(&self,
fields: Option<Vec<&str>>,
limit: Option<i32>,
sort: Option<&str>)
-> Result<Vec<Language>, Error>

The Languages endpoint returns all languages available on SyntaxCenter, sorted by position by default.

fn get_language_by_permalink<'a>(&self,
language_permalink: &'a str,
fields: Option<Vec<&str>>)
-> Result<Language, Error>

Returns the language corresponding to the permalink along with its information.

fn get_language_categories<'a>(&self,
language_permalink: &'a str,
fields: Option<Vec<&str>>,
limit: Option<i32>,
sort: Option<&str>)
-> Result<Vec<Category>, Error>

The language categories endpoint returns all of the categories corresponding to the language.

fn get_language_category_concepts<'a>(&self,
language_permalink: &'a str,
category_id: i32,
fields: Option<Vec<&str>>,
limit: Option<i32>,
sort: Option<&str>)
-> Result<Vec<Concept>, Error>

The category concepts endpoint returns all concepts belonging to the category, sorted by position by default.

fn get_language_concepts<'a>(&self,
language_permalink: &'a str,
fields: Option<Vec<&str>>,
limit: Option<i32>,
sort: Option<&str>)
-> Result<Vec<Concept>, Error>

The language concepts endpoint returns all concepts belonging to the language, sorted by position by default.

fn search_language_concepts<'a>(&self,
language_permalink: &'a str,
search: &'a str,
fields: Option<Vec<&str>>,
limit: Option<i32>)
-> Result<Vec<Concept>, Error>

This search language concepts endpoint returns all concepts within the specified language that matches the query.

fn get_concepts(&self,
fields: Option<Vec<&str>>,
limit: Option<i32>,
sort: Option<&str>)
-> Result<Vec<Concept>, Error>

The Concepts endpoint returns an array of concepts contained in SyntaxCenter. By default, the array is limited to 20 concepts, sorted by ID in ascending order.

fn search_concepts<'a>(&self,
search: &'a str,
fields: Option<Vec<&str>>,
limit: Option<i32>)
-> Result<Vec<Concept>, Error>

Full text search for matching concepts across the entire database.

fn get_concept_by_id(&self,
concept_id: i32,
fields: Option<Vec<&str>>)
-> Result<Concept, Error>

Returns the concept corresponding to the unique identifier.

fn get_language_concepts_by_concept_id(&self,
language_permalink: &str,
concept_permalink: &str,
fields: Option<Vec<&str>>)
-> Result<Concept, Error>

The language concepts endpoint returns all concepts belonging to the language, sorted by position by default.